[api] server fragment
서버 파일 분할 완료
