+++ server/academyMGS/urls.py
@@ -1,8 +1,11 @@
 from django.conf.urls import patterns, include, url
 from .views import *
 urlpatterns = (
-    url(r'^$', index),
-    url(r'^idcheck/', idCheck),
-    url(r'^register/', register),
-    url(r'^login/',login),
+    url(r'^$', indexPage),
+    url(r'^addstudent',  addStudentPage),
+    url(r'^idcheck', idCheckPage),
+    url(r'^login', loginPage),
+    url(r'^register', registerPage),
+    url(r'^search', searchPage),
+    url(r'^setstudent', setStudentPage),
 )
