+++ server/academyMGS/viewpages/setStudent.py
@@ -1,0 +1,6 @@
+from academyMGS.models import *
+import json
+from django.http import HttpResponse
+
+def setStudent(request):
+    return HttpResponse("don't access", content_type='application/json')
